refactor(landing page): clean up where components render
