+++ app/components/LandingPage.js
@@ -1,13 +1,33 @@
-import React from 'react'
+import React, { Component } from 'react'
 import Search from '../components/Search'
+import MultipleProductsContainer from '../containers/MultipleProductsContainer'
 
-function LandingPage (props) {
 
-  return (
-    <div>
-      <Search handleChange={props.handleChange} inputValue={props.inputValue} />
-    </div>
-  )
+class LandingPage extends Component {
+  constructor(props) {
+    super(props);
+    this.state = { inputValue: ''}
+    this.handleChange = this.handleChange.bind(this);
+  }
+
+  handleChange(evt) {
+    const inputValue = evt.target.value
+    this.setState({ inputValue })
+  }
+
+  render() {
+    const inputValue = this.state.inputValue
+    const filteredProducts = this.props.products.filter(product => product.name.match(inputValue))
+    console.log(filteredProducts);
+
+    return (
+      <div>
+        <Search handleChange={this.handleChange} inputValue={inputValue} />
+        <img src="" alt="" />
+        <MultipleProductsContainer />
+      </div>
+    )
+  }
 }
 
 export default LandingPage
